deoxys: migrate inout tests to integration tests
